requirements.txt
Flask>=0.11.1
Flask-Mail>=0.9.0
Flask-SQLAlchemy>=1.0
Flask-Script>=2.0.5
Flask-Bootstrap>=3.3.7
Flask-Login>=0.4.0
Flask-Bcrypt>=0.7.1
Flask-WTF>=0.13.1
python-dateutil
PyYAML>=3.11
docopt>=0.6.1
requests>=2.3.0
six>=1.7.3
tornado>=4.0
simplejson>=3.8.2
jsonschema>=2.5.1
latex>=0.6.4